Update original JSON server example
Delete now obsolete dedicated example
Add compilation instructions to advanced_usage
