--- examples/compiled-json-extension/.gitignore
@@ -1,6 +1,0 @@
-out
-node_modules
-client/server
-.vscode-test
-build/
-dist/
