--- examples/compiled-json-extension/.vscode/launch.json
@@ -1,25 +1,0 @@
-// A launch configuration that compiles the extension and then opens it inside a new window
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "extensionHost",
-            "request": "launch",
-            "name": "Launch Client",
-            "runtimeExecutable": "${execPath}",
-            "args": [
-                "--extensionDevelopmentPath=${workspaceRoot}"
-            ],
-            "outFiles": [
-                "${workspaceRoot}/client/out/**/*.js"
-            ],
-            "preLaunchTask": {
-                "type": "npm",
-                "script": "watch"
-            },
-            "env": {
-                "DebugTCP": "2087"
-            }
-        }
-    ]
-}
