--- examples/compiled-json-extension/.vscode/tasks.json
@@ -1,48 +1,0 @@
-{
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "type": "npm",
-            "script": "compile",
-            "group": "build",
-            "presentation": {
-                "panel": "dedicated",
-                "reveal": "never"
-            },
-            "problemMatcher": [
-                "$tsc"
-            ]
-        },
-        {
-            "label": "client",
-            "type": "npm",
-            "script": "watch",
-            "isBackground": true,
-            "group": "build",
-            "presentation": {
-                "panel": "dedicated",
-                "reveal": "never"
-            },
-            "problemMatcher": [
-                "$tsc-watch"
-            ]
-        },
-        {
-            "label": "server",
-            "group": "build",
-            "type": "shell",
-            "command": "make clean && make server"
-        },
-        {
-            "label": "Build Client+Server",
-            "group": {
-                "kind": "build",
-                "isDefault": true
-            },
-            "dependsOn": [
-                "client",
-                "server"
-            ]
-        }
-    ],
-}
