--- examples/compiled-json-extension/tslint.json
@@ -1,6 +1,0 @@
-{
-	"rules": {
-		"indent": [true, "tabs"],
-		"semicolon": [true, "always"]
-	}
-}
